Bake jump.tlb into OutDir at build; fix Code field type
The BakeJumpTimeline MSBuild target builds Tl.Bake and bakes
jump.json straight into $(OutDir) — IDE Run finds the asset next to
the executable. SoundClip.Code is ushort; JSON now passes integers.

+++ Timelines.cs
@@ -10,7 +10,7 @@ public readonly record struct JumpTrack(float Scale) : IBlend<JumpClip>
         => result = new(first.Height + (second.Height - first.Height) * factor);
 }
 
-public readonly record struct SoundClip(float Code);
+public readonly record struct SoundClip(ushort Code);
 
 public readonly record struct SoundTrack : IBlend<SoundClip>
 {
